Finished vfs_file_ref_create dup creation and reference test

+++ src/include/lib/queue.h
@@ -2,7 +2,7 @@
 #define QUEUE_H
 
 typedef struct queue_node_s {
-    int data;
+    long data;
     struct queue_node_s* next;
 } queue_node_t;
 
@@ -11,10 +11,10 @@ typedef struct queue_s {
     queue_node_t* rear;
 } queue_t;
 
-queue_node_t* queue_create_node(int data);
+queue_node_t* queue_create_node(long data);
 queue_t* queue_create_queue();
-void queue_enqueue(queue_t* queue, int data);
-int queue_dequeue(queue_t* queue);
+void queue_enqueue(queue_t* queue, long data);
+long queue_dequeue(queue_t* queue);
 void queue_display_queue(queue_t* queue);
