+++ src/lib/queue.c
@@ -2,7 +2,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-queue_node_t* queue_create_node(int data) {
+queue_node_t* queue_create_node(long data) {
     queue_node_t* new_node = (queue_node_t*)malloc(sizeof(queue_node_t));
     new_node->data = data;
     new_node->next = NULL;
@@ -15,7 +15,7 @@ queue_t* queue_create_queue() {
     return queue;
 }
 
-void queue_enqueue(queue_t* queue, int data) {
+void queue_enqueue(queue_t* queue, long data) {
     queue_node_t* new_node = queue_create_node(data);
     if (queue->rear == NULL) {
         queue->front = queue->rear = new_node;
@@ -25,7 +25,7 @@ void queue_enqueue(queue_t* queue, int data) {
     queue->rear = new_node;
 }
 
-int queue_dequeue(queue_t* queue) {
+long queue_dequeue(queue_t* queue) {
     if (queue->front == NULL) {
         printf("Queue is empty\n");
         return -1;
@@ -43,7 +43,7 @@ int queue_dequeue(queue_t* queue) {
 void queue_display_queue(queue_t* queue) {
     queue_node_t* temp = queue->front;
     while (temp != NULL) {
-        printf("%d ", temp->data);
+        printf("%ld ", temp->data);
         temp = temp->next;
     }
     printf("\n");
